물품 컨트롤러 수정

+++ grammy/src/main/java/com/grammy/myapp/domain/TB_StockVO.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 public class TB_StockVO {
 	private int stock_seq;
 	private String stock_name;
-	private double stock_weight;
+	private Integer stock_weight;
 	private String stock_date;
 	private String stock_shelfLife;
 	private int shelf_seq;
